requirements.txt
alembic==1.13.2
certifi==2024.6.2
distlib==0.3.8
filelock==3.15.4
Mako==1.3.5
MarkupSafe==2.1.5
pipenv==2024.0.1
platformdirs==4.2.2
psycopg2==2.9.9
python-dotenv==1.0.1
setuptools==70.1.1
SQLAlchemy==2.0.31
typing_extensions==4.12.2
virtualenv==20.26.3
